Refactor runeterra views into module
