+++ runeterra/views/__init__.py
@@ -1,0 +1,11 @@
+from .random import random_picker
+from .champion import champion_list, ChampionCreateView, ChampionUpdateView
+from .stats import region_stats
+
+__all__ = [
+    "random_picker",
+    "champion_list",
+    "ChampionCreateView",
+    "ChampionUpdateView",
+    "region_stats",
+]
